backend/requirements.txt
argostranslate==1.9.6
beautifulsoup4==4.13.5
fastapi==0.116.1
geopy==2.4.1
uvicorn
pymysql

haversine==2.9.0
httpx==0.28.1
loguru==0.7.3
numpy==2.3.2
overpy==0.7
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
rich==14.1.0
sentence_transformers==5.1.0
SQLAlchemy==2.0.43
streamlit==1.48.1
streamlit_js_eval==0.1.7
tqdm==4.67.1
